requirements.txt
numpy
torch
torchaudio
tqdm
scikit-learn
matplotlib
umap-learn